requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
alabaster==0.7.13
alembic==1.11.1
amqp==5.1.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.27.0
atomicwrites==1.4.1
attrs==21.4.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.9.1
bcrypt==4.0.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==22.6.0
blinker==1.6.2
bootstrap-py==1.0.2
bs4==0.0.1
build==0.10.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
conditional==1.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dj-rest-auth==4.0.0
Django==4.2.2
django-allauth==0.51.0
django-ckeditor==6.4.2
django-clone==3.0.6
django-cors-headers==3.13.0
django-dbbackup==4.0.1
django-embed-video==1.4.4
django-filter==22.1
django-js-asset==2.0.0
django-pgconnection==1.0.2
django-pgtrigger==2.4.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-redis==5.2.0
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.3.0
docopt==0.6.2
docutils==0.19
drf-yasg==1.21.4
email-validator==2.0.0.post2
env==0.1.0
environ==1.0
fastapi==0.97.0
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
filelock==3.11.0
flake8==4.0.1
Flask==2.3.2
form==0.0.1
frozenlist==1.3.3
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.2
h11==0.14.0
hiredis==2.2.3
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
hyperlink==21.0.0
identify==2.5.24
idna==3.4
imagesize==1.4.1
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
isort==5.10.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
magic-filter==1.0.9
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.6.1
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==0.4.3
nodeenv==1.8.0
npm==0.1.1
oauthlib==3.2.0
openai==0.26.4
optional-django==0.1.0
orjson==3.9.1
packaging==23.1
passlib==1.7.4
pathspec==0.9.0
pequests==0.0.1
pguard==0.4.0
phonenumbers==8.12.56
Pillow==9.2.0
pip-tools==6.13.0
pipreqs==0.4.13
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==3.3.1
progress==1.6
prompt-toolkit==3.0.30
psycopg2-binary==2.9.6
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
Pyrogram==2.0.97
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==1.0.0
python-environ==0.4.54
python-multipart==0.0.6
python3-openid==3.2.0
pytz==2022.2.1
pyxios==1.0.2
PyYAML==6.0
react==4.3.0
redis==4.3.4
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sentry-sdk==1.3.0
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.4
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.16
sqlparse==0.4.4
starlette==0.27.0
svgwrite==1.4.3
telebot==0.0.5
TelegramBotAPI==0.3.2
TgCrypto==1.2.5
tomli==2.0.1
tqdm==4.64.1
transliterate==1.10.2
Tree==0.2.4
Twisted==22.10.0
txaio==23.1.1
types-cryptography==3.3.23
typing_extensions==4.5.0
tzdata==2022.1
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
virtualenv==20.21.0
watchfiles==0.19.0
wcwidth==0.2.5
websockets==11.0.3
Werkzeug==2.3.4
whitenoise==6.2.0
wrapt==1.14.1
yarg==0.1.9
yarl==1.8.2
zope.interface==6.0
